requirements/common.txt
pip
Django>=3.2,<4.0
pytz==2018.7                                                                    
kombu==5.2.0
vine==5.0.0
billiard==3.6.4.0
redis==3.5.3
django-redis==5.0.0
tweepy==3.10.0
configparser==3.7.3
unidecode==1.0.23
#psycopg2-binary>=2.8,<2.9
psycopg2-binary==2.9.1
django-mptt
social-auth-app-django
git+https://github.com/DocTocToc/cleanerversion.git@django32 #update
pillow
Celery==5.1.2
python-decouple
dj-database-url
lxml
beautifulsoup4==4.7.1
djangorestframework
django-filter
django-crispy-forms
requests
django-registration
django-countries
django-sendmail-backend
python-gnupg
stripe>=2.48.0,<3.0
matplotlib
numpy
dry-rest-permissions
django-celery-beat
django-admin-rangefilter
django-meta
django-ajax-selects
furl
gunicorn==20.1.0
django-markdownify
django-bootstrap-modal-forms==2.2.0
pycountry
simplejson
git+https://github.com/DocTocToc/silver.git@21241e4a8f5a187b2190c3b722b9ddcca459ca5b #FKA django-silver
celery_once
python-twitter
ipaddress
pydiscourse
Faker
git+https://github.com/DocTocToc/django-invitations.git
django-modeltranslation
python-dateutil
selenium
django-constance[database]
django-reversion
django-phonenumber-field[phonenumbers]
django-taggit-serializer
django-user-visit==0.4.1
neomodel==4.0.8
#django-mama-cas # Central Authentication Service
git+https://github.com/DocTocToc/django-mama-cas.git@ac6623ac20b3a2655eae38161c83f3535ceaa524
gevent # django-mama-cas optional
defusedxml # django-mama-cas optional
django-dumpdata-one
django-postgres-fuzzycount
pyyaml
git+https://github.com/DocTocToc/django-durationwidget.git@800d9393ca2e99acb39ced8cd73dea27503423d1
django-widget-tweaks
django-leaflet
#django-geojson==3.1.0
git+https://github.com/DocTocToc/django-geojson.git@31cf507e504d4b7aec99b28d4c503cd8839185bc
matrix-nio[e2e]
django-webpush
fontawesomefree
humanize
